% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base functions.R
\name{data.org}
\alias{data.org}
\title{Dataset with manipulated (cleaned) x and y points}
\usage{
data.org(data)
}
\arguments{
\item{data}{input data}
}
\value{
cleaned dataset
}
\description{
This function manipulate x and y points according to KM curve characteristics, which are left and upward resistant. (x2 can't be less than x1, y2 can't be greater than y1)
}
\examples{
data <- data.org(data)
}
\keyword{internal}
