% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.table}
\alias{ex.table}
\title{Example Digitized Data for \code{km.table} Function}
\format{
Data frame with \code{x1},\code{y1},\code{x2},and \code{y2} columns for time \code{(x)} and survival rate \code{(y)} points for two groups
}
\usage{
ex.table
}
\description{
Data were extracted from Figure 1C (Malmberg, 2019) digitised by the developer. There are infomration for two groups (MRD- & MRD+), where \code{(x1, y1)} and \code{(x2, y2)} contain the pairs of time and survival rate points for the 1st and 2nd group, respectively.
}
\references{
\insertRef{delsing2019minimal}{extractKM}
}
\keyword{datasets}
