% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.functions.R
\name{km.cal.tab}
\alias{km.cal.tab}
\title{IPD Calculation}
\usage{
km.cal.tab(
  t.points,
  s.points,
  t.risk.T,
  n.risk.T,
  lower.T,
  upper.T,
  t.event = "NA",
  gr.number = "group"
)
}
\arguments{
\item{t.points}{vector of time to event points; they represents the x axis values marked from the KM plot using digizeit, which is greater than 0.}

\item{s.points}{vector of survival rate points; they represents the y axis values marked from the KM plot using digizeit, which ranges from 0 to 1.}

\item{t.risk.T}{vector of time points in the number at risk table from the original KM plot, which starts from zero.}

\item{n.risk.T}{vector of number at risk at each time point in the number at risk table from the original KM plot, which has the same length as t.risk.T.}

\item{lower.T}{numeric vector; number of data points at start between time intervals in the number at risk table}

\item{upper.T}{numeric vector; number of data points at end between time intervals in the number at risk table}

\item{t.event}{number of events}

\item{gr.number}{name for the group}

\item{n.t.T}{number of clicks in the group}
}
\value{
data frame with time to event, censoring, and group name information
}
\description{
This function calculate all the IPDs based on input data
}
\keyword{internal}
