% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM.functions.R
\name{km.event}
\alias{km.event}
\title{Extract Data from KM Curve When Number of Events Information Is Available}
\usage{
km.event(ssize, t.event = "NA", group.name = "group", data)
}
\arguments{
\item{ssize}{vector of patients' number at risk at time points from the table}

\item{t.event}{numeric value; total number of events}

\item{group.name}{name of treatment group}

\item{data}{data frame with 2 columns; 1st column has digitised time information and 2nd column has digitised survival rate information}
}
\value{
data frame with time to event, censoring, and group name information
}
\description{
This function derives or extracts IPD when only total number of events information are available
}
\examples{
df <- ex.event[,1:2]
df <- df[!is.na(df$x1) & !is.na(df$y1),]  # only valid data
res1 <- km.event(ssize=16, t.event=13, group.name="group1", data=df); res1

df <- ex.event[,3:4]
df <- df[!is.na(df$x2) & !is.na(df$y2),]
res2 <- km.event(ssize=39, t.event=28, group.name="group2", data=df); res2

library(survival)
result <- rbind(res1,res2); 
km <- survfit(Surv(Surv.Time, Censor) ~ Group, data=result);km
plot(km)
}
\references{
\insertRef{guyot2012enhanced}{extractKM}
}
