% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM.functions.R
\name{km.tab.event}
\alias{km.tab.event}
\title{Extract Data from KM Curve When Both Number at Risk Table and Number of Events Are Available}
\usage{
km.tab.event(t.risk.T, n.risk.T, t.event = "NA", group.name = "group", data)
}
\arguments{
\item{t.risk.T}{vector of time points from the number at risk table}

\item{n.risk.T}{vector of patients' number at risk at time points from the table}

\item{t.event}{numeric value; total number of events}

\item{group.name}{name of treatment group}

\item{data}{data frame with 2 columns; 1st column has digitised time information and 2nd column has digitised survival rate information}
}
\value{
data frame with time to event, censoring, and group name information
}
\description{
This function derives or extracts IPD when both number at risk table and total number of events information are available
}
\examples{
df <- ex.tab.event[,1:2]
df <- df[!is.na(df$x1) & !is.na(df$y1),]  # only valid data
res1 <- km.tab.event(t.risk.T = c(0,24,48,72,96,120), 
                     n.risk.T = c(23,11,8,4,2,0), 
                     t.event=5, 
                     group.name="group1",
                     data=df);res1

df <- ex.tab.event[,3:4]
df <- df[!is.na(df$x2) & !is.na(df$y2),]
res2 <- km.tab.event(t.risk.T = c(0,24,48,72,96,120),
                     n.risk.T = c(12,3,3,2,1,1),
                     t.event=8,
                     group.name="group2",
                     data=df);res2

library(survival)
result <- rbind(res1,res2); 
km <- survfit(Surv(Surv.Time, Censor) ~ Group, data=result);km
plot(km)
}
\references{
\insertRef{guyot2012enhanced}{extractKM}
}
