% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM.functions.R
\name{km.table}
\alias{km.table}
\title{Extract Data from KM Curve When Number at Risk Table Is Available}
\usage{
km.table(t.risk.T, n.risk.T, group.name = "group", data)
}
\arguments{
\item{t.risk.T}{vector of time points from the number at risk table}

\item{n.risk.T}{vector of patients' number at risk at time points from the table}

\item{group.name}{name of treatment group}

\item{data}{data frame with 2 columns; 1st column has digitised time information and 2nd column has digitised survival rate information}
}
\value{
data frame with time to event, censoring, and group name information
}
\description{
This function derives or extracts IPD when only number at risk table information are available
}
\examples{
df <- ex.table[,1:2]
df <- df[!is.na(df$x1) & !is.na(df$y1),]  # only valid data
res1 <- km.table(t.risk.T = c(0,25,50,75,100,125),
                 n.risk.T = c(22,18,10,7,5,2),
                 group.name="group1",
                 data=df);res1

df <- ex.event[,3:4]
df <- df[!is.na(df$x2) & !is.na(df$y2),]
res2 <- km.table(t.risk.T = c(0,25,50,75),
                 n.risk.T = c(5,1,1,0),
                 group.name="group2",
                 data=df);res2

library(survival)
result <- rbind(res1,res2); 
km <- survfit(Surv(Surv.Time, Censor) ~ Group, data=result);km
plot(km)
}
\references{
\insertRef{guyot2012enhanced}{extractKM}
}
